backend/requirements.txt
'''
Authors: Lukas Pietzschmann, Vincent Ugrai
'''

flask
flask-restful
pymongo
dnspython
flask_cors
python-dotenv
requests
bs4
websockets
azure-storage-blob
